added config package to keep track of any functions for accessing config file
